refactor: NewNotificationEventHandler 패키지 수정
